backend/requirements.txt
# ============================================
# Transport MVP - Backend Dependencies
# ============================================

# Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Validation & Serialization
pydantic[email]==2.10.4
pydantic-settings==2.7.1

# Auth
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Firebase
firebase-admin==6.6.0

# HTTP Client
httpx==0.28.1

# Dev
python-dotenv==1.0.1
